Test delivering reminders in development
Add letter_opener gem to see emails that are sent by the application in
development.

Also add a rake task to create past due reminders to ease testing in
development. I had to change the reminder factory to dynamically
evaluate the due_date field for the task to work. Otherwise it would
create the reminder with the time that the factory was loaded.